refactor: separate to Canvas and tool-bar
