+++ src/Canvas/assets/canvas.css
@@ -1,0 +1,1 @@
+.Canvas {}
