+++ src/Canvas/assets/tool-bar.css
@@ -1,0 +1,10 @@
+.tool-bar ul {
+  list-style: none;
+  position: fixed;
+  left: 5px;
+  top: 45%;
+}
+
+.tool-bar ul li {
+  cursor: pointer;
+}
