+++ src/Canvas/index.js
@@ -1,0 +1,4 @@
+import Canvas from './canvas';
+import ToolBar from './tool-bar';
+
+export { Canvas, ToolBar };
