+++ src/demo/assets/demo.css
@@ -1,0 +1,5 @@
+.Demo {
+  transition: all 0.3s cubic-bezier(0.645, 0.045, 0.355, 1);
+  background: pink;
+  height: 100vh;
+}
