--- src/mock/tools.json
@@ -1,18 +1,0 @@
-[
-  {
-    "key": "minus",
-    "value": "line",
-    "isBind": true
-  },
-  {
-    "key": "border",
-    "value": "border",
-    "style": {
-      "width": 100,
-      "height": 80,
-      "background": "#F96",
-      "borderRadius": "10px",
-      "border": "1px solid #aaa"
-    }
-  }
-]
