+++ src/options/tools.js
@@ -1,0 +1,20 @@
+// key: 用作 Icon 的 type
+// value: 唯一标识
+// style: tool 最外层样式
+export const tools = [
+  {
+    key: 'minus',
+    value: 'line',
+  },
+  {
+    key: 'border',
+    value: 'block',
+    style: {
+      width: 100,
+      height: 80,
+      background: '#F96',
+      borderRadius: '10px',
+      border: '1px solid #aaa',
+    },
+  },
+];
