--- src/urlHelper/path.js
@@ -1,11 +1,0 @@
-const proxy = null;
-
-const dev = {
-  tools: '../../mock/tools.json',
-};
-
-const prod = {};
-
-const path = dev;
-
-export { path, proxy };
